ai-service/requirements.txt
pydantic-ai==0.0.14
pydantic==2.10.3
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.0
python-multipart==0.0.12
httpx==0.28.1
supabase==2.10.0
openai==1.58.1